feat: added preview menu item to kustomization and helm-vaues files in file explorer
